[NXP][src] Add support for TBR management cluster and secondary nwk, Updated common app to work with SNI and TBRM
