+++ src/platform/nxp/rt/rw61x/PlatformManagerImpl.cpp
@@ -204,6 +204,7 @@ CHIP_ERROR PlatformManagerImpl::_InitChipStack(void)
     otPlatLogInit();
     otPlatRadioInit();
     otPlatSetResetFunction(initiateResetInIdle);
+    otPlatRandomInit();
 #endif
 
 #if CHIP_DEVICE_CONFIG_ENABLE_WPA
